refactor: controller에서 재입력을 받도록 로직 변경
